requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.49.0
h11==0.14.0
idna==3.6
imutils==0.5.4
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pip==24.0
pycparser==2.21
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
selenium==4.18.1
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.42.0
wsproto==1.2.0